+++ src/components/Header/Header.jsx
@@ -37,18 +37,18 @@ export default function Header() {
       </div>
 
       {isLogged ? (
-        <div className={`flex-row-wrap flex-mid-center gap20`}>
+        <div className={`${styles.links} flex-row-wrap flex-mid-center gap20`}>
           <Link
             to="/"
             onClick={logoutHandler}
-            className="links  btn btn_action subtitle-1"
+            className={`btn btn_action subtitle-1`}
           >
             Logout
           </Link>
         </div>
       ) : (
-        <div className={`flex-row-wrap flex-mid-center gap20`}>
-          <Link to="/login" className="links btn btn_action subtitle-1">
+        <div className={`${styles.links} flex-row-wrap flex-mid-center gap20`}>
+          <Link to="/login" className={`btn btn_action subtitle-1`}>
             Sign In
           </Link>
           <Link to="sign-up" className="links btn btn_action subtitle-1">
