+++ src/components/Header/Header.module.css
@@ -6,7 +6,6 @@
   height: 100%;
   background-color: var(--BG);
   color: var(--TEXT);
-  justify-content: space-between;
   align-items: center;
   position: sticky;
   top: 0;
@@ -31,6 +30,10 @@
 .userAvatar {
   background-color: transparent;
   border: none;
+}
+
+.links {
+  margin-left: auto;
 }
 
 .Avatar_options {
