+++ src/pages/Explore/Explore.jsx
@@ -20,6 +20,14 @@ export default function Explore() {
         sidebar ? styles.openSesame : ""
       } flex-column-wrap flex-mid-center content`}
     >
+      <div className={styles.page_header}>
+        <h1>Explore</h1>
+
+        <h4 className="subtitle-1">
+          {/* Total Videos :{" "}
+          {findPlaylist.videos?.length ? findPlaylist.videos?.length : 0} */}
+        </h4>
+      </div>
       <CategoryChips />
       <div className={`${styles.videos_container} `}>
         {filteredVideos?.map((video) => (
