+++ src/pages/Explore/Explore.module.css
@@ -28,6 +28,19 @@
   transition: background-color var(--ANIM) 300ms;
 }
 
+.page_header {
+  height: fit-content;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 2rem;
+}
+
+.trash_btn {
+  margin-left: auto;
+}
+
 .openSesame {
   filter: blur(4px);
 }
