+++ src/pages/History/History.module.css
@@ -7,6 +7,15 @@
   color: var(--TEXT);
 }
 
+.page_header {
+  height: fit-content;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 2rem;
+}
+
 .emptyImage {
   margin-top: 4rem;
   width: 24rem;
@@ -26,6 +35,28 @@
   display: none;
   color: var(--TEXT);
   transition: filter, display var(--ANIM) 1000ms;
+}
+
+.dialog {
+  gap: 10px;
+  align-items: center;
+  justify-content: flex-start;
+  outline: 2px solid;
+  margin-bottom: 1rem;
+  border-radius: 8px;
+  font-size: small;
+  font-weight: bold;
+  background-color: inherit;
+  transition: background-color var(--ANIM) 300ms;
+}
+
+.dialog:hover {
+  background-color: var(--HOVER_BG_ACTIVE);
+  transition: background-color var(--ANIM) 300ms;
+}
+
+.trash_btn {
+  margin-left: auto;
 }
 
 @media only screen and (max-width: 479px) {
