+++ src/pages/Liked/Liked.module.css
@@ -28,6 +28,19 @@
   transition: filter, display var(--ANIM) 1000ms;
 }
 
+.page_header {
+  height: fit-content;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 2rem;
+}
+
+.trash_btn {
+  margin-left: auto;
+}
+
 @media only screen and (max-width: 479px) {
   .emptyImage {
     margin-top: 3rem;
